Rename image to fix change
