+++ src/layouts/InterfaceLayout/containers/DashboardContainer/DashboardContainer.vue
@@ -42,7 +42,7 @@
                 <span
                   :class="['currency-symbol', 'cc', pair.from, 'cc-icon']"
                 ></span>
-                <img src="@/assets/images/icons/swap.svg" alt />
+                <img src="@/assets/images/icons/swap-widget.svg" alt />
                 <span
                   :class="['currency-symbol', 'cc', pair.to, 'cc-icon']"
                 ></span>
